pyproject.toml
[project]
name = "airqualitycast"
version = "0.1.0"
description = "AirQualityCast — production data pipeline for hyperlocal AQI forecasting in India."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "tenacity>=9.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "psycopg[binary,pool]>=3.2",
    "pandas>=2.2",
    "pyarrow>=17.0",
    "structlog>=24.4",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "types-python-dateutil",
]
eda = [
    "jupyter",
    "matplotlib",
    "seaborn",
    "folium",
    "plotly",
    "sqlalchemy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airqualitycast"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "airflow/dags", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "N"]
"airflow/dags/**" = ["B"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["psycopg.*", "pandas.*", "pyarrow.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --cov=src/airqualitycast --cov-report=term-missing"
asyncio_mode = "auto"
markers = [
    "integration: requires a running TimescaleDB",
]
